# Add unroll pragma for Clang in XXH3_accumulate.