Clang doesn't unroll the XXH3_accumulate loop for some reason. Using `#pragma clang loop unroll(enable)` to hint to Clang that it should unroll results in a huge 1.4-1.5x speedup.

Before: 
```
XXH3_64bits         :     102400 ->   162910 it/s (15909.2 MB/s)
XXH3_64b unaligned  :     102400 ->   140762 it/s (13746.3 MB/s)
```
After:
```
XXH3_64bits         :     102400 ->   215992 it/s (21093.0 MB/s)
XXH3_64b unaligned  :     102400 ->   188791 it/s (18436.6 MB/s)
```