@Cyan4973 2019-03-09
Wow ! That's a great speed up ! Congratulation @easyaspi314 !

Also : note that I had to change the formula a little bit today.
Further analysis shows that UMAC is flawed for checksumming purpose,
but thankfully, it takes only a little modification to overcome this deficiency.

The new code is <5% slower, due to additional operation in the critical loop.
I haven't updated the ARM code, I know this part much less.