@easyaspi314 2019-03-09
Okay…well that'll make things a lot more difficult on ARM. I would have to reshuffle, because ARM does this:
```c
uint32_t xdata[4] = { A, B, C, D };

uint32x2x2_t d = { { A, C }, { B, D } };
```

At this point, ARM is stuck with `d` shuffled.

I guess I could do `vshrn`/`vmovn` instead, as that might be the better option at this point, but as you saw in my last PR, that slows things down on non-Cortex-A57s.