@Cyan4973 2019-03-09
Just FYI : 
the `#pragma clang loop unroll` slows things down on my ARM setup (Qualcom chip, clang 7.0, termux).
Non-turbo speed goes down from 5.4 GB/s to 5.1 GB/s.
Not a big issue, and not sure why.
I have no doubt it increases speed for you.
It just shows once again how hard it is to find optimizations that work positively for all ARM variants.