@easyaspi314 2019-03-09
Hmm. Well, at this point, ARM is back to the drawing board. I haven't tested the unroll on ARM yet.

Side note: Scalar code looks like it will break if there is a carry.