@Cyan4973 2019-03-09
oh, you mean, the `#pragma` speeds up `clang` for x64 ?

I haven't noticed any difference on my setup,
but then, `clang` on my laptop and desktop is really good, 
so I don't think there is any performance left on the table.
note: I'm using `clang 7.0`

I'm sorry the modification make the ARM code tricky.
This addition is necessary to guarantee that all input constituents will be present in the final mix.
Without it, there is a probability that a 4-byte segment get just ignored (it's nullified).
Even though it happens on average once every 16 GB, which may seem rare, that's still way too much for a 64+ bits hash, and for the purpose of checksumming.

Adding the input was the simplest additional operation I could find.
In spite of this, even on x86/x64, it can result in a visible drop, depending on CPU and compiler, since it's right in the critical loop. For comparison, I get up to 20% less bandwidth on my desktop, while the laptop is unaffected, and the ARM phone is <5% slower.