@easyaspi314 2019-03-09
Well how does this do?
```c
_mm_add_epi32(xacc[i], _mm_shuffle_epi32(d, _MM_SHUFFLE(2, 3, 0, 1)));
```

Because I could definitely work with that, as I can just combine and use `umull2` on NEON. 

If not, I can just do `vshrn`/`vmovn`, but that means 6-7 more cycles in the main loop.