@Cyan4973 2019-03-09
I tried something like that initially (though I xored, instead of add).
It would reduce performance by >30%, which was too impactful.