@easyaspi314 2019-03-09
Before on my phone (it's not firing on all cores)
```
XXH3_64bits         :     102400 ->    80447 it/s ( 7856.2 MB/s)
XXH3_64b unaligned  :     102400 ->    80244 it/s ( 7836.4 MB/s)
```
After:
```
XXH3_64bits         :     102400 ->    59310 it/s ( 5792.0 MB/s)
XXH3_64b unaligned  :     102400 ->    59674 it/s ( 5827.5 MB/s)
```