@easyaspi314 2019-03-09
I can get it a little faster on my phone (6.2 GB/s) with a different shuffling method and by adding beforehand, however, it becomes incredibly slow on the A57:
```
./xxhsum 0.6.6 (64-bits little endian), by Yann Collet
Sample of 100 KB...
XXH32               :     102400 ->    37991 it/s ( 3710.1 MB/s)
XXH32 unaligned     :     102400 ->    37819 it/s ( 3693.3 MB/s)
XXH64               :     102400 ->    25980 it/s ( 2537.1 MB/s)
XXH64 unaligned     :     102400 ->    25969 it/s ( 2536.0 MB/s)
XXH3_64bits         :     102400 ->    39652 it/s ( 3872.3 MB/s)
XXH3_64b unaligned  :     102400 ->    38851 it/s ( 3794.0 MB/s)
```

Fortunately, it just looks like terrible codegen from Clang 3.8. I am going to try cross compiling with newer versions of Clang and GCC.