@easyaspi314 2019-03-09
Cross compiled in a Manjaro VM with both Clang 7.0.1 and aarch64-linux-gnu-gcc 8.2.0
```
~$ ./xxhsum.clang7 -b
./xxhsum.clang7 0.6.6 (64-bits little endian), by Yann Collet
Sample of 100 KB...
XXH32               :     102400 ->    38211 it/s ( 3731.5 MB/s)
XXH32 unaligned     :     102400 ->    37852 it/s ( 3696.5 MB/s)
XXH64               :     102400 ->    25979 it/s ( 2537.0 MB/s)
XXH64 unaligned     :     102400 ->    25974 it/s ( 2536.6 MB/s)
XXH3_64bits         :     102400 ->    50487 it/s ( 4930.4 MB/s)
XXH3_64b unaligned  :     102400 ->    49648 it/s ( 4848.4 MB/s)
~$ ./xxhsum.gcc8 -b
./xxhsum.gcc8 0.6.6 (64-bits little endian), by Yann Collet
Sample of 100 KB...
XXH32               :     102400 ->    38221 it/s ( 3732.5 MB/s)
XXH32 unaligned     :     102400 ->    37857 it/s ( 3696.9 MB/s)
XXH64               :     102400 ->    25994 it/s ( 2538.5 MB/s)
XXH64 unaligned     :     102400 ->    25986 it/s ( 2537.7 MB/s)
XXH3_64bits         :     102400 ->    39240 it/s ( 3832.1 MB/s)
XXH3_64b unaligned  :     102400 ->    39031 it/s ( 3811.6 MB/s)
```

5GB/s is much better.

Using `vshrn`/`vmovn` is faster like I saw before, but considering what happened before, that might not be the same in your case:
```
~$ ./xxhsum.clang7.shrn -b
./xxhsum.clang7.shrn 0.6.6 (64-bits little endian), by Yann Collet
Sample of 100 KB...
XXH32               :     102400 ->    38223 it/s ( 3732.7 MB/s)
XXH32 unaligned     :     102400 ->    37880 it/s ( 3699.2 MB/s)
XXH64               :     102400 ->    25981 it/s ( 2537.2 MB/s)
XXH64 unaligned     :     102400 ->    25976 it/s ( 2536.7 MB/s)
XXH3_64bits         :     102400 ->    57480 it/s ( 5613.2 MB/s)
XXH3_64b unaligned  :     102400 ->    55917 it/s ( 5460.6 MB/s)
~$ ./xxhsum.gcc8.shrn -b
./xxhsum.gcc8.shrn 0.6.6 (64-bits little endian), by Yann Collet
Sample of 100 KB...
XXH32               :     102400 ->    38197 it/s ( 3730.1 MB/s)
XXH32 unaligned     :     102400 ->    37838 it/s ( 3695.1 MB/s)
XXH64               :     102400 ->    25987 it/s ( 2537.8 MB/s)
XXH64 unaligned     :     102400 ->    25982 it/s ( 2537.3 MB/s)
XXH3_64bits         :     102400 ->    48220 it/s ( 4709.0 MB/s)
XXH3_64b unaligned  :     102400 ->    47879 it/s ( 4675.7 MB/s)
```

Also, I just realized I was doing a terrible life algorithm again.
Smart me: Use the vbox shared folder to copy my ssh key over and scp that from the vm.
Actual me: Build on manjaro, copy to the vbox shared folder on my Mac, and scp from there. 🤦‍♂️ 

I put the changes in the branch called `xxh3-aarch64-test`. If you define `ALT`, it will use the `vzip` method, otherwise it will use the `vshrn`/`vmovn` method. Could you please test it on your S7?