@Cyan4973 2019-03-10
Here are some results from S7 with `clang7` under `termux` : 
```
XXH3_64bits() : 
before : 5.7 GB/s
after, without ALT : 4.8 GB/s
after, with -DALT : 4.8 GB/s
```
I couldn't measure a significant difference between the two variants.

For comparison, `XXH64` is at `3.8 GB/s` on this system.