@easyaspi314 2019-03-10
I checked the assembly output on ARMv7 from Clang and I noticed that was making copies before the vzip. I modify it in place with an inline assembly string and I get ~7.2. 

Me:
![](https://i.imgur.com/8jj1cZW.png)
enough