@Cyan4973 2019-03-11
I've started adding some self-tests within `xxhsum`.
They are triggered every time the benchmark command `-b` is invoked.

These hash values are validated vs the `SSE2` path, which is the most important one.
They are supposed to help us find potential discrepancies between targets and code paths (AVX2, NEON and scalar).