@easyaspi314 2019-03-12
Side note: The scalar version bus errors on ARMv7. Alignment memes are still an issue on ARM, mainly with the `ldrd` instruction. I switch to `XXH_read32/64` and it works