@easyaspi314 2019-03-12
I have access to mips64, ppc64le, ppc64, sun4u, and sparc64 in the GCC compiler farm. You might want to make an account because they'll let you use it for testing and building as long as it is open source software. https://cfarm.tetaneutral.net/

Additionally, while SSE slows down XXH32 and XXH64 on x86_64. with NEON intrinsics, my G3 loves it and is about 2 times faster (2.2 GB/s->4.1 GB/s) on XXH32 and about 3 times faster (0.8 GB/s->2.4 GB/s) on XXH64. The A57 doesn't get a huge difference though, mainly because of the native (albeit slow) 64-bit integers and the higher NEON latency.